requirements.txt
dDjango==5.0.6
Pillow==8.4.0
torch
transformers
nltk
requests
